backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
PyJWT==2.8.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
Werkzeug==3.0.1
